requirements.txt
# Core requirements
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.0
PyYAML>=5.4.1
tqdm>=4.61.0
scikit-learn>=0.24.2
boto3>=1.17.78
opencv-python>=4.5.3
requests>=2.25.1

# Testing and development tools (optional)
pytest>=6.2.5
flake8>=3.9.2
black>=21.6b0